+++ python/scratch_python_project.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+
+cookiecutter "${HOME}/.dotfiles/python/cookiecutter/scratch-python-template" timestamp="$(date +%Y%m%d%H%M)"